Updated balancer APY and price implementations for Polygon
